Reworked logic of _backup_dir & added in further backup checks

+++ installation_and_upgrade/ibex_install_utils/tasks/backup_tasks.py
@@ -18,19 +18,21 @@ class BackupTasks(BaseTasks):
             self.prompt.prompt_and_raise_if_not_yes(
                 f"You appear to be trying to delete the folder, {src}, containing the current working directory {os.getcwd()}. "
                 f"Please do this manually to be on the safe side")
-        elif os.path.exists(backup_dir):
+            return
+        if os.path.exists(backup_dir):
             self.prompt.prompt_and_raise_if_not_yes(
                 f"Backup dir {backup_dir} already exists. Please backup this app manually")
-        elif os.path.exists(src):
+            return
+        if os.path.exists(src):
             if copy:
                 print(f"Copying {src} to {backup_dir}")
                 shutil.copytree(src, backup_dir, ignore=ignore)
             else:
                 print(f"Moving {src} to {backup_dir}")
                 self._file_utils.move_dir(src, backup_dir, self.prompt)
-        else:
-            self.prompt.prompt_and_raise_if_not_yes(f"You appear to be running a backup without first installing IBEX. Please manually check your installation first.")
-            
+        else: #if src can't be found on the machine
+            self.prompt.prompt_and_raise_if_not_yes(f"You appear to backing up {src}, but it doesn't exist on this machine. If it is Python, this is likely okay. Please manually check your installation first.")
+            return
 
     @task("Backup old directories")
     def backup_old_directories(self):
@@ -79,19 +81,22 @@ class BackupTasks(BaseTasks):
         GUI_PATH_BACKUP = os.path.join(self._get_backup_dir(),'Client_E4')  
         SETTINGS_PATH = os.path.join(INSTRUMENT_BASE_DIR, "Settings")
         AUTOSAVE_PATH = os.path.join(INSTRUMENT_BASE_DIR, "Var", "Autosave")
+        UTILS_PATH = os.path.join(self._get_backup_dir(), 'EPICS_UTILS')
 
         backup_paths = (EPICS_PATH_BACKUP, PYTHON3_PATH_BACKUP, GUI_PATH_BACKUP)
 
         for d in backup_paths:
-            if !(os.path.exists(os.path.join(d, "VERSION.txt"))): 
+            if not(os.path.exists(os.path.join(d, "VERSION.txt"))):
                 backup_failed_str = "Backup failed at " + d
                 self.prompt.prompt_and_raise_if_not_yes(f(backup_failed_str))
                 self.prompt.prompt_and_raise_if_not_yes(f"Error found with backup. Continue? ")
 
-        if !os.path.exists(SETTINGS_PATH):
+        if not(os.path.exists(SETTINGS_PATH)):
             self.prompt.prompt_and_raise_if_not_yes(f"Error found with backup. Settings did not back up properly. Continue? ")
-        if !os.path.exists(AUTOSAVE_PATH):
+        if not(os.path.exists(AUTOSAVE_PATH)):
             self.prompt.prompt_and_raise_if_not_yes(f"Error found with backup. Autosave did not back up properly. Continue? ")
+        if not(os.path.exists(UTILS_PATH)):
+            self.prompt.prompt_and_raise_if_not_yes(f"Error found with backup. EPICS_Utils did not back up properly. Continue? ")
